requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
